+++ dynts/media/dynts/ecoplot/ecoplot.js
@@ -308,179 +308,6 @@
     
 }(jQuery));
 
-/**
- * Scatter plot plugin for flot.
- * It allows to have different color for points referring to different dates
- */
-(function($) {
-    
-    $.plot.plugins.push((function() {        
-        var defaults = {
-            series: {
-                scatter: {
-                    radius: 5,
-                    lineWidth: 1,
-                    gradient: 1,
-                    symbol: 'circle',
-                }
-            }
-        };
-        
-        function calculateColors(plot, series) {
-            var opts = plot.getOptions(),
-                legend = opts.legend;
-        }
-        
-        // Get the colour scheme for each point
-        function scatter_colours(plot, series, datapoints) {
-            if(!series.scatter){return;}
-            var opts = plot.getOptions(),
-                data = series.data,
-                NC = opts.colors.length,
-                N = data.length,
-                S = Math.floor(N/(NC-1)),
-                scatter = opts.series.scatter,
-                colors = [],
-                minv,maxv,c,lv,dc,sdata,i;
-            if(scatter.data) {return;}
-            scatter.data = sdata = [];
-            if(series.lines) {series.lines.show = false;}
-            if(series.points) {series.points.show = false;}
-            if(series.bars) {series.bars.show = false;}
-            if(!N) {return;}
-            scatter.colors = colors;
-            if(S) {
-                for(i=1;i<NC;i++) {
-                    if(i == NC-1) {
-                        S = N - S*(NC-2);
-                    }
-                    $.each($.color_gradient(opts.colors[i-1],
-                                            opts.colors[i],
-                                            S,
-                                            scatter.gradient),function(i,val) {
-                        colors.push(val);
-                    });
-                }
-            }
-            minv = maxv = data[0][2];
-            $.each(data,function(i,v) {
-                lv = v[2];
-                minv = Math.min(minv,lv);
-                maxv = Math.max(maxv,lv);
-            });
-            $.each(data,function(i,v) {
-                lv = v[2];
-                c = (lv - minv)/(maxv - minv);
-                sdata.push(Math.floor(c*N));
-                if(lv === maxv) {scatter.last = i;}
-            });
-        }
-        
-        function gradient_div(colors) {
-            var el = $('<div>').css({'padding':'1px'}),
-                i,c0,c1;
-            for(i=1;i<colors.length;i++) {
-                c0 = colors[i-1];
-                c1 = colors[i];
-                $('<div>').width(10).height(20).appendTo(el)
-                        .css({'background-color':c1,
-                              'background-image': '-moz-linear-gradient(top, '+c1+','+c0+')',
-                              'background-image': '-webkit-linear-gradient('+c1+','+c0+')',
-                              'background-image': 'linear-gradient(top, '+c1+','+c0+')'});
-            }
-            return el;
-        }
-        
-        function drawSeries(plot, ctx, series) {
-            var options = plot.getOptions(),
-                scatter = options.series.scatter,
-                colors = scatter.colors,
-                plotOffset = plot.getPlotOffset(),
-                points = options.series.scatter,
-                sw = options.series.shadowSize,
-                lw = points.lineWidth,
-                radius = points.radius,
-                symbol = points.symbol,
-                gradient = gradient_div(options.colors),
-                self = plot.getPlaceholder(),
-                legend = $('.legend',self),
-                table = $('table',legend); 
-            
-            if(!scatter) {return;}
-            
-            $('.legendColorBox',table).empty().append(gradient);
-            legend.remove();
-            legend = $('<div class="legend">').prependTo(self);
-            legend.append($('<table>').html(table.html()));
-            legend.draggable({containment:self});
-            
-            function fillcolor(i) {
-                return colors[scatter.data[i]];
-            }
-            
-            function plotPoints(datapoints, radius, fillStyle, offset, shadow,
-                                axisx, axisy, symbol) {
-                var points = datapoints.points, ps = datapoints.pointsize;
-                
-                for(var i = 0; i < points.length; i += ps) {
-                    var x = points[i], y = points[i + 1];
-                    if (x == null || x < axisx.min || x > axisx.max || y < axisy.min || y > axisy.max)
-                        continue;
-                    
-                    ctx.beginPath();
-                    x = axisx.p2c(x);
-                    y = axisy.p2c(y) + offset;
-                    if (symbol == "circle")
-                        ctx.arc(x, y, radius, 0, shadow ? Math.PI : Math.PI * 2, false);
-                    else
-                        symbol(ctx, x, y, radius, shadow);
-                    ctx.closePath();
-                    
-                    if(fillStyle) {
-                        ctx.fillStyle = fillcolor(i);
-                        ctx.fill();
-                    }
-                    ctx.stroke();
-                }
-            }
-            
-            ctx.save();
-            ctx.translate(plotOffset.left, plotOffset.top);
-
-            if (lw > 0 && sw > 0) {
-                // draw shadow in two steps
-                var w = sw / 2;
-                ctx.lineWidth = w;
-                ctx.strokeStyle = "rgba(0,0,0,0.1)";
-                plotPoints(series.datapoints, radius, null, w + w/2, true,
-                           series.xaxis, series.yaxis, symbol);
-                ctx.strokeStyle = "rgba(0,0,0,0.2)";
-                plotPoints(series.datapoints, radius, null, w/2, true,
-                           series.xaxis, series.yaxis, symbol);
-            }
-
-            ctx.lineWidth = lw;
-            ctx.strokeStyle = series.color;
-            plotPoints(series.datapoints, radius, fillcolor, 0, false,
-                       series.xaxis, series.yaxis, symbol);
-            ctx.restore();
-        }
-
-        function _init(plot) {
-            plot.hooks.processDatapoints.push(scatter_colours);
-            plot.hooks.drawSeries.push(drawSeries);
-        }
-        
-        return {
-            name: 'scatter',
-            version: '0.1',
-            init:_init,
-            options: defaults
-        };
-    }()));
-    
-}(jQuery));
-
 
 (function($) {
 
@@ -555,6 +382,22 @@
                        b.click(function(e) {
                            $this.trigger('load',[$this, this]);
                        });
+                   }
+               },
+               {
+                   classname: 'toggle-legend',
+                   title: "Toggle legend",
+                   icon: "ui-icon-comment",
+                   type: "checkbox",
+                   decorate: function(b,el) {
+                       var $this = $(el);
+                       b.toggle(function() {
+                            $('.legend',el).hide();
+                        },
+                        function() {
+                            $('.legend',el).show();
+                        }
+                      );
                    }
                },
                {
@@ -733,16 +576,16 @@
          */
         function _editpanel(data, showplot, oldcanvas) {
             // check if oldcanvas is the same. If so keep it!
-            var newcanvas = data;
-            var table = null;
-            var body  = null;
-            var oldbody = null;
-            var oseries = [];
+            var newcanvas = data,
+                table = null,
+                body  = null,
+                oldbody = null,
+                oseries = [];
             if(!oldcanvas) {
                 $.ecoplot.log.debug('Creating editing panel.');
                 table = $('<table class="plot-options"></table>');
                 var head = $('<tr></tr>').appendTo($('<thead></thead>').appendTo(table));
-                head.html('<th>serie</th><th>line</th><th>points</th><th>bars</th><th>y-axis1</th><th>y-axis2</th>');
+                head.html('<th>line</th><th>points</th><th>bars</th><th>y-axis1</th><th>y-axis2</th>');
                 body = $('<tbody></tbody>').appendTo(table);
                 table.click(function() {
                     data.render();
@@ -798,7 +641,6 @@
                 }
                 var trt = $('<tr class="line'+circle+' serie'+i+' serie-title"></tr>').appendTo(body);
                 var tr  = $('<tr class="line'+circle+' serie'+i+' serie-option"></tr>').appendTo(body);
-                tr.append($('<td></td>'));
                 trt.append($('<td class="label" colspan="6">'+serie.label+'</td>'));
                 tr.append(tdinp('checkbox','line','line', serie.lines.show));
                 tr.append(tdinp('checkbox','points','points', serie.points.show));
@@ -898,7 +740,7 @@
         };
 
         /**
-         * Internal function for setting up one or more flot plot depending on data.
+         * Internal function for setting up one or more new flot plot depending on data.
          * It creates the options.canvases object of the form:
          * 
          * options.canvases = {
@@ -906,6 +748,9 @@
          *      height: height of canvas
          *      current: index of current canvas or null
          *      }
+         * 
+         * @param $this jQuery ecoplot element
+         * @params data Json data
          */
         function _set_new_canavases($this,data) {
             var options = $this[0].options;
@@ -920,7 +765,7 @@
             options.canvases.all = canvases;
 
             var optpanel = function(c) {
-                // create options pnel if not available
+                // create options panel if not available
                 var cn = 'option'+c;
                 var opt = $('.secondary .panel.'+cn,$this);
                 if(!opt.length) {
@@ -1146,16 +991,16 @@
                 }
                 var ax = pl.flot.getAxes();
                 var opts = {};
-                if(ax.xaxis.used)  {
+                if(ax.xaxis)  {
                     opts.xaxis = checkax(ranges.xaxis);
                 }
-                if(ax.yaxis.used)  {
+                if(ax.yaxis)  {
                     opts.yaxis = checkax(ranges.yaxis);
                 }
-                if(ax.x2axis.used)  {
+                if(ax.x2axis)  {
                     opts.x2axis = checkax(ranges.x2axis);
                 }
-                if(ax.y2axis.used)  {
+                if(ax.y2axis)  {
                     opts.y2axis = checkax(ranges.y2axis);
                 }
                 // do the zooming