+++ dynts/__init__.py
@@ -1,6 +1,6 @@
 '''Quantitative financial timeseries analysis'''
 
-VERSION = (0, 4, 1)
+VERSION = (0, 4, 2)
  
 def get_version():
     return '.'.join(map(str,VERSION))