+++ dynts/backends/tsnumpy.py
@@ -8,6 +8,7 @@ import numpy as np
 
 import dynts
 from dynts import lib, composename
+from dynts.conf import settings
 from dynts.utils import laggeddates, asarray
 
 
@@ -23,7 +24,8 @@ _functions = {'min':'min',
 
 
 def rollsingle(self, func, window = 20, name = None,
-               fallback = False, **kwargs):
+               fallback = False, align = 'right',
+               **kwargs):
     '''Efficient rolling window calculation for min, max type functions'''
     rname = 'roll_{0}'.format(func)
     if fallback:
@@ -36,7 +38,12 @@ def rollsingle(self, func, window = 20, name = None,
     data = np.array([rolling(serie) for serie in self.series()])
     name = name or self.makename(func,window=window)
     dates = asarray(self.dates())
-    return self.clone(dates[window-1:], data.transpose(), name = name)
+    desc = settings.desc
+    if (align == 'right' and not desc) or desc:
+        dates = dates[window-1:]
+    else:
+        dates = dates[:-window+1]
+    return self.clone(dates, data.transpose(), name = name)
     
 
 def days(d1,d0):
@@ -61,7 +68,10 @@ class TimeSeries(dynts.TimeSeries):
     
     @property
     def shape(self):
-        return self._data.shape
+        if self._data is not None:
+            return self._data.shape
+        else:
+            return (0,0)
     
     def __getitem__(self, i):
         return self._data[i]