+++ dynts/conf.py
@@ -2,40 +2,54 @@ from math import isnan
 toupper = lambda x : str(x).upper()
 
 class Settings(object):
-    '''The setting class contains configuration parameters used in dynts.
+    '''\
+The setting class contains configuration parameters used in dynts.
+
+.. attribute:: backend
+
+    Default :class:`dynts.TimeSeries` backend class.
     
-    .. attribute:: backend
+    Default ``"numpy"``.
+
+.. attribute:: desc
+
+    Default sorting order for dates. If ``False`` timeseries are ordered in 
+    ascending date order, if ``True`` in a descending order.
     
-        the default :class:`dynts.TimeSeries` backend class, Default ``zoo``.
-        
-    .. attribute:: concat_operator
+    Default ``False``.
     
-        The operator for concatenating expressions. Default ``,``.
+.. attribute:: concat_operator
+
+    Operator for concatenating expressions.
     
-    .. attribute:: default_loader
+    Default ``,``.
+
+.. attribute:: default_loader
+
+    Default :class:`dynts.data.TimeSerieLoader` class.
+    If this is ``None`` the :class:`dynts.data.TimeSerieLoader` class will be used as loder.
     
-        Default :class:`dynts.data.TimeSerieLoader` class. Default ``None``.
-        If this is ``None`` the :class:`dynts.data.TimeSerieLoader` class will be used as loder.
-        
-    .. attribute:: default_provider
+    Default ``None``.
     
-        Default :class:`dynts.data.DataProvider` code. Default ``"YAHOO"``.
-        
-    .. attribute:: field_separator
+.. attribute:: default_provider
+
+    Default :class:`dynts.data.DataProvider` code. Default ``"YAHOO"``.
     
-        Character used to separate tickers from fields and providers. Default ``:``.
-        
-    .. attribute:: months_history
+.. attribute:: field_separator
+
+    Character used to separate tickers from fields and providers. Default ``:``.
     
-        the default number of months of history. Default: ``12``.
-        
-    .. attribute:: proxies
+.. attribute:: months_history
+
+    the default number of months of history. Default: ``12``.
     
-        dictionary of proxy servers. Default ``{}``.
-        If you need to use a proxy server to access the web::
-        
-            from dynts.conf import settings
-            settings.proxies['http'] = 'http://yourproxy.com:80'
+.. attribute:: proxies
+
+    dictionary of proxy servers. Default ``{}``.
+    If you need to use a proxy server to access the web::
+    
+        from dynts.conf import settings
+        settings.proxies['http'] = 'http://yourproxy.com:80'
                 
 To change settings::
     
@@ -45,6 +59,7 @@ To change settings::
     '''
     def __init__(self):
         self.backend = 'numpy'
+        self.desc = False
         self.splittingnames     = '__'
         self.concat_operator    = ','
         self.separator_operator = '|'