+++ dynts/web/__init__.py
@@ -1,0 +1,3 @@
+'''Web views and utilities for djpcms
+https://github.com/lsbardel/djpcms
+'''